I'm kinda surprised I need to SvREFCNT_inc the return value ...
But I suppose it makes sense since I bump it back to perl in a
lexical that goes out of scope.  Apparently the XS can touch
things without touching them so my av_fetch() doesn't inc the
refcount even when it's pushed into the lexical...  if I said
that right.